Bug #777753: GCC variation with link-time optimization (LTO)
When GCC does link-time optimizations, it embeds some randomness in
the generated object.  This is https://bugs.debian.org/777753, and it
affects several packages.

it may be the underlying cause of several instances of
build_id_variation_requiring_further_investigation, but i haven't
removed those issues yet, because there may be other build_id
variation mechanisms on the packages affected by LTO as well, and i
haven't tested them all.

The changes to packages.yml were found based on the "affects:"
annotations on #777753.