Drop "gettext_creates_ChangeLog_files_and_entries_with_current_date" issue.
It is actually gettextize who creates changelog files and entries, but
gettextize is not a tool to be run at build time anyway (that's what
autopoint is for), so it's probably better to deal with this on
a per package basis (there was only one package affected).

Move bug number from issues.yml to packages.yml.